Working direct install flow
